PDF convert will throw a specific conversion_error when conversion fails

+++ app/api/upload/specs/PDF.spec.js
@@ -139,7 +139,12 @@ describe('PDF', function () {
     });
 
     afterEach((done) => {
-      fs.unlink(pdf.optimizedPath, done);
+      if (pdf.optimizedPath) {
+        fs.unlink(pdf.optimizedPath, done);
+      }
+      else {
+        done();
+      }
     });
 
     it('should optimize and extract html and text', (done) => {
@@ -152,6 +157,20 @@ describe('PDF', function () {
         done();
       })
       .catch(done.fail);
+    });
+
+    describe('when there is a conversion error', () => {
+      it('should throw a conversion_error', (done) => {
+        spyOn(pdf, 'optimize').and.returnValue(Promise.reject());
+        pdf.convert()
+        .then(() => {
+          done.fail('should have thrown a conversion_error');
+        })
+        .catch(error => {
+          expect(error).toEqual({error: 'conversion_error'});
+          done();
+        });
+      });
     });
   });
 });
